fix size overflow in C implementation of radix sort

+++ c_algorithms/radixsort.c
@@ -13,8 +13,8 @@ int get_max_value(uint32_t arr[], int n) {
 }
  
 // Counting Sort based on exp
-void countSort(uint32_t arr[], int n, int exp) {
-    int output[n];
+void countSort(uint32_t arr[], int n, uint64_t exp) {
+    uint32_t output[n];
     int i, count[10] = { 0 };
  
     // Store count of occurrences in count[]
@@ -33,14 +33,15 @@ void countSort(uint32_t arr[], int n, int exp) {
     // Copy the output array to arr[]
     for (i = 0; i < n; i++)
         arr[i] = output[i];
+
 }
 
 uint32_t *radix_sort(uint32_t *arr, size_t size) {
     // Find max value to know max number of digits
-    int m = get_max_value(arr, size);
- 
+    uint32_t m = get_max_value(arr, size);
+
     // Perform counting sort on each digit
-    for (int exp = 1; m / exp > 0; exp *= 10)
+    for (uint64_t exp = 1; m / exp > 0; exp *= 10)
         countSort(arr, size, exp);
 
     return arr;
